moved read function to dirutil
