+++ fileutil/fileutil.go
@@ -1,29 +1,12 @@
 package fileutil
 
 import (
-	"fmt"
 	"github.com/bijeshos/goose/dirutil"
 	"io"
 	"log"
 	"os"
 	"path/filepath"
 )
-
-//Read to read from dir
-func Read(srcDir string) []string {
-	fmt.Println("reading files from: ", srcDir)
-	var files []string
-	err := filepath.Walk(srcDir, func(path string, info os.FileInfo, err error) error {
-		if !info.IsDir() {
-			files = append(files, path)
-		}
-		return nil
-	})
-	if err != nil {
-		panic(err)
-	}
-	return files
-}
 
 func CopyFile(srcPath, targetPath string) {
 	//open source file
